Prevent file deletion on backend.
The 'delete' button on a file upload field is still there, but only
deletes the file reference, and not the file itself. This prevents
crashes when people press 'delete' on the file and then 'cancel' on
the node edit form. It also allows for sharing of files.

The file itself might become an orphan, which we should handle
independent from this issue.